requirements/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
matplotlib 
tqdm
